Fix first essay missing from CSV
When fetching the essays, the Python script would output

```
✅ 001 This Year We Can End the Death Penalty in California
❌ 001 This Year We Can End the Death Penalty in California, ('list' object has no attribute 'keys')
```

It succeeded at downloading the 1st essay in Markdown, but it would fail
when writing it to the CSV file. This is because a DictWriter was being
used to write the CSV, and it expects a dictionary for each row, but it
was being passed a list such as

`[1, 'This Year We Can End the Death Penalty in California', '2016-01-01', 'http://www.paulgraham.com/prop62.html']`

We could convert the list to a dict but it's simpler to use a plain csv
writer since the row lines have their values in columnar order.