Rename merge target to markdown
We merge the essays for all different formats of the book. That is not
unique to the markdown version of the book. Name the target after the
book format, not the process of merging the essays which they all have
in common.

It's possible that the intention was to use the single generated .md file
as the input for the other book formats, but that did not happen.