Stop ignoring essays and certain book formats
It will be useful to have essays under version control, as we can see
diffs or have archives if an essay ever gets retracted.

As far not ignoring certain book formats, we already have some formats
staged, not sure why others were being excluded, seems a mistake. The
epub book must have been force staged since it was being ignored.

Eventually I want to move to having a build directory where the final
artifacts (the various book formats) are not staged. This is because
they are auto-generated and will update frequently, not a good candidate
for version control. Readers can instead download the books from the
Releases page and easily access older versions of the books too. That
means the build directory (or its contents rather) will be ignored soon.