Add individual essays in Markdown
Except for 3 and 4.

Staging them allows us to track changes as essays can have updates, and
gives us a backup as essays can also be unpublished.

It also opens the possibility for optimising the build. For example,
treat this as a cache with a long TTL and then on most builds you don't
need to scrape all 200+ essays, just fetch the new ones, which should
shave off quite a bit of time relative to the build time.