Fix broken links for external essays
Essays 3 and 4 are hosted on a different website, which means they are
absolute links instead of relative links unlike all the other essays.

Aaron's RSS feed only handles relative links by prefixing the string
"http://www.paulgraham.com/" which breaks the absolute links.

The proper fix is to switch to a more robust RSS feed. I had a quick
search and didn't find any other feeds, so we'll need to make one.

At that point we can also fix other issues at the source, such as
publication dates which we try to handle with htmldate but that still
gets them very wrong because the essays have no date info on the page.