Exit with error code on fetch failure
This way the caller (e.g. CI build) can decide to handle failures, such
as aborting instead of trying to build an incomplete or even empty book.

And print a list of the essays we failed to fetch. May get rid of this
output or put it behind an optional verbose flag, since we already have
inline essay status on each fetch iteration (but that doesn't group the
failures together for easier reading).